Restructured project by separating the Program function into separate functions. Added ConsoleMenus folder to hold all the menus. Moved askQuestion into a new file under common called ConsoleFunctions.js
