+++ src/ConsoleApp.js
@@ -1,29 +1,13 @@
-const readline = require('readline');
-const path = require("path");
-const { StudentDataReader, TeacherDataReader } = require("./DataLayer");
-const { StudentService } = require("./Services");
-const { Student, Teacher } = require("./Models");
-
-const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-});
-
-function askQuestion(question) {
-    return new Promise((resolve, reject) => {
-        rl.question(question, (ans) => {
-            resolve(ans);
-        })
-    });
-}
-
+const { askQuestion } = require("./Common/ConsoleFunctions");
+const StudentMenu = require("./ConsoleMenus/StudentMenu")
 
 async function Program() {
-    const baseFilePath = path.join(__dirname, "../", "JSONData");
-    const _studentDataReader = new StudentDataReader(path.join(baseFilePath, "Students.json"));
-    const _teacherDataReader = new TeacherDataReader(path.join(baseFilePath, "Teachers.json"));
-    const _studentService = new StudentService(_studentDataReader, _teacherDataReader);
+    // Welcome Message
+    await MainMenu();
+    // Thank you message
+}
 
+async function MainMenu() {
     let shouldLoop = true;
     while (shouldLoop) {
         console.log("[1] Students");
@@ -32,38 +16,7 @@ async function Program() {
         let userInput = await askQuestion("Select an option from above: ");
         switch (userInput) {
             case "1":
-                console.log("[1] Add Student");
-                console.log("[2] Search Students");
-                console.log("[3] Update Student");
-                console.log("[4] Delete Student");
-                console.log("[5] Go Back");
-                let userInputStudent = await askQuestion("Select an option from above: ");
-                switch (userInputStudent) {
-                    case "1":
-                        let studentFirstName = await askQuestion("Enter Student First Name: ");
-                        let studentLastName = await askQuestion("Enter Student Last Name: ");
-                        let studentAge = await askQuestion("Enter Student Age: ");
-                        let parsedStudentAge = parseInt(studentAge);
-                        let grades = await askQuestion("Enter Student Grades (Space-Separated): ");
-                        let teacherId = await askQuestion("Enter Teacher's ID: ");
-                        let parsedGrades = grades.split(" ").map(num => parseInt(num));
-                        let newStudent = new Student(
-                            studentFirstName,
-                            studentLastName,
-                            parsedStudentAge,
-                            parsedGrades,
-                            teacherId
-                        );
-                        _studentService.addStudent(newStudent);
-                        break;
-                    case "2":
-                        let searchTerm = await askQuestion("Enter search term: ");
-                        let matchingStudents = _studentService.searchByName(searchTerm);
-                        console.log(matchingStudents);
-                        break;
-                    default:
-                        console.log("Going back to main menu");
-                }
+                await StudentMenu();
                 break;
             case "2":
                 break;
